Split up events tests into separate files
